+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## v0.0.3 - 2025-09-19
 - Initalize npm package for development
+- Create and implement Drawer Counter Web Component
 ## 0.0.2 - 2025-09-19
 - Add initial project files including HTML, CSS, JS, and service worker for PWA functionality
 - Adds web components for banners, toasts, network status, install and launch app buttons